graph saved to pickle file
